requirements.txt
django
djangorestframework
gunicorn
dj-database-url
PyMySQL
django-cors-headers
whitenoise
pytz
openpyxl
